// src/enums.ts
export enum dgramTypes {
  UDP4 = "udp4",
  UDP6 = "udp6",
}

export enum availability {
  Availabile = "availabile",
  Unavailabile = "unavailabile",
}

export enum defaultOptions {
  MULTICAST_ADDRESS = "224.0.0.234",
  DEFAULT_UDP_PORT = 44201,
  DEFAULT_TIMEOUT = 1000,
  DEFAULT_INTERVAL = 3000,
  GLOBAL_EVENT_NAME = "MessageBus",
}